Restructure to prepare for 2nd lambda
